[Silver III] Title: 이친수, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/2193. 이친수/이친수.cc
@@ -5,15 +5,13 @@
 using namespace std;
 
 int main() {
-    int n;
+    long long n;
     cin >> n;
     
-    // 데이터 타입을 long long으로 변경
-    vector<vector<long long>> dp(n+1, vector<long long>(2));
+    vector<vector<long long>> dp(n+1,vector<long long>(2));
     // 이친수는 1이 두번 나타나지 않음
-    // dp[i][j] = i길이의 j로 끝나는 이친수의 개수
-    // dp[i][0] = dp[i-1][0] + dp[i-1][1]
-    // dp[i][1] = dp[i-1][0]
+    // dp[i][j] = i길이의 j로 끝나는 이친수 의 개수 = dp[i][0] = dp[i-1][0] + dp[i-1][1] 
+    // dp[i][1] = dp[i-1][0] 
 
     dp[1][0] = 0;
     dp[1][1] = 1;
@@ -22,7 +20,6 @@ int main() {
         dp[i][1] = dp[i - 1][0];
     }
 
-    // 출력 타입도 long long으로
     cout << dp[n][0] + dp[n][1] << '\n';
 
     return 0;
